Support user signup
Include the beginnings of the Backend class which will wrap web API
functionality behind a C# interface. Also includes the start of
reasonable error handling. To be fair, the API's handling of errors is
pretty ad-hoc so there isn't much we can do, but I'm looking to
standardize some error strings soon which should help us handle it
better here.